Factor out `default` modules
