--- bot/http/handler/default/mod.ts
@@ -1,1 +1,0 @@
-export * from "./handler.ts";
