+++ bot/http/handler/handler_interface.ts
@@ -1,0 +1,6 @@
+/**
+ * HandlerInterface for HTTP requests.
+ */
+export interface HandlerInterface {
+  handle: (req: Request) => Promise<Response>;
+}
