+++ bot/http/handler/mod.ts
@@ -1,1 +1,2 @@
+export * from "./handler_interface.ts";
 export * from "./handler.ts";
